portfolio_tracker/requirements.txt
nsepython
requests
python-dotenv
